refactor: 우승상금 클래스 이름 변경 및 2등 확인 필드 추가
