+++ src/main/java/me/daeho/lotto/domain/Rank.java
@@ -1,0 +1,43 @@
+package me.daeho.lotto.domain;
+
+
+import java.util.Arrays;
+
+public enum Rank {
+    NONE(0, 0, false),
+    FIFTH(3, 5000, false),
+    FOURTH(4, 50000, false),
+    THIRD(5, 1500000, false),
+    SECOND(5, 3000000, true),
+    FIRST(6,2000000000, false)
+    ;
+
+    private final int count;
+    private final int prize;
+    private final boolean matchBonus;
+
+    Rank(int count, int prize, boolean matchBonus) {
+        this.count = count;
+        this.prize = prize;
+        this.matchBonus = matchBonus;
+    }
+
+    public int getCount() {
+        return count;
+    }
+
+    public int getPrize() {
+        return prize;
+    }
+
+    public boolean isMatchBonus() {
+        return matchBonus;
+    }
+
+    public static Rank valueOf(int count, boolean matchBonus) {
+        return Arrays.stream(values())
+                .filter(v -> count == v.count && matchBonus == v.matchBonus)
+                .findAny()
+                .orElse(NONE);
+    }
+}
